+++ CHANGELOG.md
@@ -7,4 +7,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - sends KeepAlive pings until report is sent
-- Adds zipper module for zipping reports directory
+- Adds zipper module for zipping html-report directory
+- Adds sender module for posting html-report archives to `httpserver`